IE will sometimes report an focusOffset greater than expected
If you control-click to the right of the final section in a mobiledoc,
IE's `getSelection()` reports that the editor is the focusNode and the
focusOffset will be equal to the length of its childNodes.
This adds a method `endOf` to the selection-utils that finds the end of
the given node, descending its `lastChild` until it finds one with no
children.